requirements.txt
selenium #@latest
requests #@latest
python-dotenv #@latest
locust #@latest